# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AMUPF Digital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department (President only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  # --- USERS / KYC ---
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List Peace Forum members
      parameters:
        - name: department_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: position
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Add a student as a Peace Forum member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    patch:
      operationId: updateMember
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /members/link/{uniqueLink}:
    get:
      operationId: getMemberByLink
      tags: [members]
      summary: Get member by unique profile link
      parameters:
        - name: uniqueLink
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "403":
          description: Access denied
        "404":
          description: Not found
  /members/{id}/regenerate-link:
    post:
      operationId: regenerateMemberLink
      tags: [members]
      summary: Regenerate a member's unique profile link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New link generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # --- TASKS ---
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registered

  # --- ATTENDANCE ---
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        executive_id:
          type: ["string", "null"]
        executive_name:
          type: ["string", "null"]
        member_count:
          type: integer
        report_count:
          type: integer
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        executive_id:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        executive_id:
          type: string

    # --- REPORT ---
...
    Member:
      type: object
      required: [member_id, user_id, position, department_id, join_date, unique_link, is_active, created_at]
      properties:
        member_id:
          type: string
        user_id:
          type: string
        full_name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        profile_photo:
          type: ["string", "null"]
        position:
          type: string
          enum: [member, secretary, coordinator, vice_president, president]
        department_id:
          type: integer
        department_name:
          type: ["string", "null"]
        join_date:
          type: string
        unique_link:
          type: string
        link_expiry:
          type: ["string", "null"]
        responsibilities:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
    MemberInput:
      type: object
      required: [user_id, position, department_id, join_date]
      properties:
        user_id:
          type: string
        position:
          type: string
          enum: [member, secretary, coordinator, vice_president, president]
        department_id:
          type: integer
        join_date:
          type: string
        responsibilities:
          type: string
        link_expiry:
          type: string
    MemberUpdate:
      type: object
      properties:
        position:
          type: string
        department_id:
          type: integer
        responsibilities:
          type: string
        is_active:
          type: boolean
        link_expiry:
          type: string

    # --- TASK ---
...
    Event:
      type: object
      required: [id, name, event_date, location, created_by, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        event_date:
          type: string
        location:
          type: string
        max_attendees:
          type: ["integer", "null"]
        target_audience:
          type: string
          enum: [all, members, executives]
        created_by:
          type: string
        created_by_name:
          type: ["string", "null"]
        attendee_count:
          type: integer
        created_at:
          type: string
    EventInput:
      type: object
      required: [name, event_date, location, target_audience]
      properties:
        name:
          type: string
        description:
          type: string
        event_date:
          type: string
        location:
          type: string
        max_attendees:
          type: integer
        target_audience:
          type: string
          enum: [all, members, executives]
    EventUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        event_date:
          type: string
        location:
          type: string
        max_attendees:
          type: integer

    # --- ATTENDANCE ---
...