# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMUPF Digital Management System API
 * OpenAPI spec version: 1.0.0
 */

export interface MemberUpdate {
  position?: string;
  department_id?: number;
  responsibilities?: string;
  is_active?: boolean;
  link_expiry?: string;
}